adding more to mcp
